+++ app/src/main/java/edu/up/cs/androidcatan/GameState.java
@@ -2,10 +2,16 @@ package edu.up.cs.androidcatan;
 
 import android.util.Log;
 import android.widget.EditText;
+
+import edu.up.cs.androidcatan.catan.buildings.Building;
+import edu.up.cs.androidcatan.catan.buildings.City;
+import edu.up.cs.androidcatan.catan.buildings.Road;
+import edu.up.cs.androidcatan.catan.buildings.Settlement;
+import edu.up.cs.androidcatan.catan.devcards.DevelopmentCard;
+import edu.up.cs.androidcatan.catan.hexagon.Hexagon;
 import gameframework.*;
 import java.util.ArrayList;
 import java.util.Random;
-import java.util.Set;
 
 /**
  * @author Alex Weininger
