+++ app/src/main/java/edu/up/cs/androidcatan/catan/buildings/Building.java
@@ -1,4 +1,4 @@
-package edu.up.cs.androidcatan;
+package edu.up.cs.androidcatan.catan.buildings;
 
 /**
  * @author Alex Weininger
