+++ app/src/main/java/edu/up/cs/androidcatan/catan/buildings/City.java
@@ -1,6 +1,8 @@
-package edu.up.cs.androidcatan;
+package edu.up.cs.androidcatan.catan.buildings;
 
 import java.util.Arrays;
+
+import edu.up.cs.androidcatan.catan.buildings.Building;
 
 /**
  * @author Alex Weininger
@@ -10,26 +12,27 @@ import java.util.Arrays;
  * @version October 30th, 2018
  * https://github.com/alexweininger/android-catan
  **/
-public class Settlement extends Building {
 
-    final static int[] resourceCost = {1, 1, 1, 0, 1}; // Brick, Grain, Lumber, Ore, Wool
+public class City extends Building {
+
+    static final int[] resourceCost = {0, 2, 0, 3, 0}; // Brick, Grain, Lumber, Ore, Wool
 
     /**
-     * @param ownerId - player id of who owns the settlement
+     * @param ownerId id of who owns the building
      */
-    Settlement(int ownerId) {
+    City(int intersectionID, int ownerId) {
         super(ownerId);
-    } // end constructor
-
-    public int getVictoryPoints() {
-        return 1;
     }
 
+    @Override
+    public int getVictoryPoints() {
+        return 2;
+    }
 
     @Override
     public String toString() {
-        return "Settlement{" +
+        return "City{" +
                 "resourceCost=" + Arrays.toString(resourceCost) +
                 '}';
     }
-} // end Class
+}
