+++ app/src/main/java/edu/up/cs/androidcatan/catan/buildings/Settlement.java
@@ -1,7 +1,8 @@
-package edu.up.cs.androidcatan;
+package edu.up.cs.androidcatan.catan.buildings;
 
 import java.util.Arrays;
-import java.util.HashMap;
+
+import edu.up.cs.androidcatan.catan.buildings.Building;
 
 /**
  * @author Alex Weininger
@@ -11,27 +12,26 @@ import java.util.HashMap;
  * @version October 30th, 2018
  * https://github.com/alexweininger/android-catan
  **/
+public class Settlement extends Building {
 
-public class City extends Building {
-
-    static final int[] resourceCost = {0, 2, 0, 3, 0}; // Brick, Grain, Lumber, Ore, Wool
+    final static int[] resourceCost = {1, 1, 1, 0, 1}; // Brick, Grain, Lumber, Ore, Wool
 
     /**
-     * @param ownerId id of who owns the building
+     * @param ownerId - player id of who owns the settlement
      */
-    City(int intersectionID, int ownerId) {
+    Settlement(int ownerId) {
         super(ownerId);
+    } // end constructor
+
+    public int getVictoryPoints() {
+        return 1;
     }
 
-    @Override
-    public int getVictoryPoints() {
-        return 2;
-    }
 
     @Override
     public String toString() {
-        return "City{" +
+        return "Settlement{" +
                 "resourceCost=" + Arrays.toString(resourceCost) +
                 '}';
     }
-}
+} // end Class
