+++ app/src/main/java/edu/up/cs/androidcatan/catan/graphics/House.java
@@ -1,4 +1,4 @@
-package edu.up.cs.androidcatan;
+package edu.up.cs.androidcatan.catan.graphics;
 
 import android.graphics.Canvas;
 import android.graphics.Color;
